+++ jeomo-mem/src/main/java/com/jeomo/mem/service/impl/MemberLevelServiceImpl.java
@@ -1,8 +1,8 @@
 package com.jeomo.mem.service.impl;
 
-import com.jeomo.mem.service.IMemberLevelService;
-import io.swagger.models.auth.In;
 import org.springframework.stereotype.Service;
+
+import com.jeomo.mem.service.IMemberLevelService;
 
 /**
  * @Author: qbt
@@ -13,7 +13,7 @@ import org.springframework.stereotype.Service;
 public class MemberLevelServiceImpl implements IMemberLevelService {
 
     @Override
-    public Integer queryDefaultMemberLevel() {
-        return 1;
+    public Long queryDefaultMemberLevel() {
+        return 1L;
     }
 }
